vending_pc/requirements.txt
pyserial-asyncio==0.6
PyQt6==6.7.0
qasync==0.27.1
